This is a big one with a lot of changes I just didn't commit due to laziness.
Fixed an issue with the new text handling where the texture would still be marked as loaded when setting the text. There were a multitude of issues occurring around this such as the texture being filled in with a random texture, the opengl texture name being unloaded immediately after being created, and the texture just never being created in the first place.

Changed the button UI component to use the new hover handling.

Changed how the mouse cursor is handled to be a UIObject instead of just being a random object getting rendered with the original render function. The biggest upsides of this are that I can place the cursor more accurately with the SetPositionFromAnchor method instead of having to place the point of the cursor in the center of the texture as well as being able to easily change animations on different events.

Added some new handling regarding dragging. It's been changed to be event based as well as having a preview event where the drag can be previewed and optionally accepted or declined based on any arbitrary calculations you might want to do. The mouse delta position has also been added as a parameter to dragging so it is now possible to put a drag event on an object and use the delta values from that to move a completely unrelated object without affecting the object you are dragging.

Changed how the cursor object was handled to now be a static object inherent to the game window instead of requiring it to be passed down the chain to every object that might need to use it.

Changed hover and hover end events to not be run in tasks.

Changed the UIObject's ReverseTree object to just hold UIObjects instead of UITreeNodes as the extra info is no longer necessary with the current handling.

Moved all of the instanced render data into its own folder and created TileInstancedRenderData as well as a new set of tile shaders to handle rendering tiles.

Added UIInstancedRenderData and some shaders for it but they haven't been fully implemented yet and the old handling is still being used.

Added a DrawGame function to the RenderFunctions object. This allows UI rendering and game rendering to be easily separated if that's something that will be necessary for future post processing.

Moved game ticking out of the render thread. At this point it's assumed that any time an opengl operation needs to be run it will be queued into the render cycle to be handled.

Completely changed how tilemaps are managed and rendered. Each tile is now a 3D tile object with an associated tile pillar (to show differences in height). Tilemaps are now loaded when the camera moves instead of via a button click and a larger amount are loaded into memory at a time but only a handful are rendered at any given time. This makes for a much more seamless experience when playing although some optimizations can still be done (such as only assigning a tile pillar to a tile where the underside could potentially be visible).

Fixed the cursor's sensitivity being tied to the update speed.

Added a specific culling function for units and added a check for whether the tilemap they are in is currently visible. This saves a lot of unnecessary calculations when there are a lot of units outside of the visible tilemaps.

Added a "calculation thread" to handle expensive and frequent calculations that shouldn't spawn a new thread every time they need to be done. The only use right now is vision map updates as they are too expensive to be run synchronously and run frequently enough that thread starvation is a real concern. Some further optimizations to vision calculation (only updating vision for the unit that moves instead of doing a full recalculation for the entire team) are currently on the agenda but I haven't gotten there yet.

Changed camera update handling to use events instead of weird sporadically placed functions and methods.

Added some non-sickening camera movements. Currently this has 16 rotation options and 5 vertical angle options.

Added a dedicated "Definitions" folder to handle static declarations of things in engine.

Added a TemplateRangedSingleTarget ability that can be inherited to easily create simple single target effects.

Added an ApplyPassives and RemovePassives method to abilities so that abilities can manage the passive effects they provide to units.

Changed ability trees to not have defined parent and child relationships to other nodes and instead now just have a connection to another node. The placement of the nodes will be decided by me on the ability tree ui.

Added a way to track loaded entities in the entity manager.

Added some toggles to the game footer to switch between normal abilities, item abilities, and consumable abilities.

Removed the background from unit status bars.

Added items, inventory, and unit equipment. The inventory contains information for all items it contains and how much gold it has. This inventory can be serialized to make saving it easier.
Items are non serializable but item entries contain all of the information necessary to recreate them. Items can be equipped to units assuming they contain the proper equipment slot definition and the unit can equip items in that slot.

Added a "bounding square" calculation to feature bounds to cut down on expensive calculations when many bounding points are loaded. This bounding square encloses the maximum points reached by the bounding points and so its a quick and dirty calculation to see if a point lies within an area where a full calculation might be necessary.

Added a PlayerParty object. This contains the units that are members of the main party and provides some handling for grouping and ungrouping them. This is also where the inventory information for the player will be stored.

Added an alternate version of DeserializableDictionary that uses a separate namespace. This is the only way I could think to fix nesting issues.

Added all the new handling to the save state.

Greatly simplified the information required for UnitSaveInfo by making UnitInfo and UnitAI serializable.

Fixed camera movement being kinda janky when multiple keys were held down. Now the camera position is only updated once regardless of how many keys are held down.

Added a quest log and ability tree UI. Inventory is next on the agenda.

Simplified quest completion checks.

Moved most tile handling out of the TileMapController and into TileMapHelpers or TileMapManager.

Added a new UnitSpritesheet

Removed texture mixing from the default fragment shader