Added event actions. These will be used as the primary method of interactivity in the game and essentially act as arbitrary code that can be made in the tools and run in game. Each event action will need to be manually created but once it is made it shouldn't need to ever be touched up.
Fixed a stack overflow error caused by maxing out the available sound buffers.

Changed the text font to arial and added some settings to make the text crispier.

Changed icons to not use multiple base objects

Added a constructor to UIBlock to create a block with a list of animations.

Added some UI handling for focused objects (ie objects that will capture all user interaction while they exist)

Added box selecting for units and heavily reworked unit selection to not be so jank.

Completely reworked vision to now store vision as groups of sets/dictionaries of tiles in vision instead of creating a massive array and filling it with vision data.
Vision is also now calculated via tilemap line calculations again but tweaked a bit from the old naive version to ensure full coverage (in this iteration we get lines to create one sextant of a hexagon and then rotate that sextant 5 times to complete the hexagon. This means that as long as full coverage exists for the initial sextant we will have full coverage of the entire hexagon)

Added some handling to automatically dispose of textures every once in a while. The current method allows 500 texture handles to build up which then get emptied in one batch. This may seem like an excessive amount of textures to save up before emptying but it's not too crazy when you factor in that every text string has a texture handle associated with it.

Took a stab at getting UI instanced rendering working but ended up giving up on that. Maybe I'll give it another shot at some point if performance becomes a big concern.

Changed ability name and description fields from strings to text info for localization purposes.

Added some group abilities that become available when multiple controlled units are selected.

Added a template ability for aoes. This can be inherited by another ability and have most of the functionality required to have an aoe targeted ability right out of the box.

Reworked abilities to use reflection to create instances of the ability instead of weird hand rolled creation functions.

Reworked tile effects. Instead of assigning tile effects directly to tiles they are now passed through a TileEffectManager static class that associates tile effects with tile points via dictionary.
Added a sample WeakSpiderWeb tile effect.

Removed the BasicAbility distinction from abilities. Any handling that was going to be inherent to "basic abilities" will just be handled through other avenues.

Simplified some handling in the GameFooter because it was getting quite confusing to look through and work with.

Added some branching paths to the update footer function based on the FooterMode enum. Right now the supported footer modes are single unit and multi unit. Single unit will display the abilities the the selected unit contains and multi unit will display the relevant group abilities.

Made several small tweaks and additions to item/inventory/equipment handling. Changed the ItemManager to automatically gather all item defitions via reflection instead of requiring hand-made Items.Add() calls for each and every item.

Added a UnitItemEntry class to simplify the assignment of items to units from the engine tools.

Added Inventory and Equipment UIs.

Created static manager classes for animation sets, dialogue, spritesheets, and unitinfo. These simplify the loading of unloaded data and fetching of already loaded data. The managers also support grouping the data into blocks of data of certain sizes so that we won't run into cases where we are loading info for 20k units when we just need to access 1, for example.

Added a feature units class to features. These act exactly the same as an affected point but come with an attached UnitId which simplifies feature creation and handling.

Added QuestRewards. These can be gold, items, or lua scripts.

Added an ID_LEDGER static class. This stores the highest id for each relevant type in a file and simplifies retrieving new ids.

Reworked BaseTiles to remove the AttachedTile and UnitOnTile fields. The AttachedTile field was being used to lazily store relations between tiles which was being monopolized almost entirely by the move ability to store selection tiles and the UnitOnTile field was causing issues with phased movement where a unit could walk over a phased unit and disassociate it from its tile.

The AttachedTile handling has been replaced by improved handling of selection tiles in the TileMapController and the UnitOnTile handling has been replaced by a UnitPositionManager static class that tracks the positions of units (each unit still contains a reference to its TileMapPosition for simplicity's sake)

Removed all non-selection tile handling from the TileMapController class as tile maps are now being handled by the TileMapManager.

Added a more nuanced check for whether to create tile pillars which ensures that the tile has a height higher than a neighboring tile before giving it a pillar.

Added a rewards section to the quest log.

Added a scale field to units.
Added an "innate tile offset" field to units that is calculated based on the scale and X rotation of the unit. This innate offset is used to place the unit in the center of their tile without requiring the information to be entered manually for each unit. (the tile offset and selection tile offset fields still exist and can be used in conjunction with the innate offset)

Moved UnitInfo into its own file.

Fixed a veeeeeeery old bug in fastDefaultShader where columns and rows were being reversed when calculating the spritesheet position. This took so long to find due to the issue only occuring with non-square spritesheets.