Improved performance when loading tile maps. Specifically, time was gained by running operations that had no bearing on each other in tasks as well as resolving map brushes in groups of 20 in tasks instead of one at a time.
Fixed the Transformations2D class applying the transformations in the incorrect order.
Added an option to scale the matrix from an arbitrary anchor point.

Replaced unit selection tiles with a new SelectionIndicatorManager static class. This allows all specific unit selection tile handling to be moved out of units and into the manager. When a unit is selected they now request a selection indiactor from the manager, who then takes one from it's object pool and fills all the necessary info for the unit.

Changed how unit selection tiles work model-wise to now use an individual mesh. This mesh is created from the tile the unit is standing on as well as all of the neighboring tiles. The texture for the selection indicator is then scaled down and placed under the unit. When the unit's position changes (and not the tile map position), the texture is moved accordingly to follow the unit. When the unit's tile map position changes, the individual mesh is recreated with the new neighboring tiles.

Modified the rendering process to now support deferred shading. Deferred shading is used by default for terrain, units, and structures while generic objects and low priority objects still use immediate shading.
The advantage of deferred shading is that many more lights can be applied to objects using deferred shading which isn't too useful right now but having the option available should be nice in the future.

Added a way to specify the desired texture wrapping to the Texture class.

Changed the default camera angle to be shallower.

Added some randomness to the throw dirt animation via new FloatRange and IntRange structs. These allow a range of values to be specified as well as provide a simple way to generate a value in that range.

Fixed an issue where buff effects wouldn't be copied over into a new dictionary when using a prototype buff. This resulted in the prototype buff having its buff effects cleared when the first child buff was removed from a unit.

Fixed an issue where features weren't checking for a ledgered unit before spawning a new unit of the same permanent id.

Added an UpdateFromDirectBitmap method to textures to avoid needing to lock and unlock the bits when updating.