Added several new CubeMethods functions to support getting cube positions via float coordinates.
Added a perlin noise generator function.

Added a new Override and OverrideContainer set of objects that provides a way to specify a list of object overrides with built in priorities. For example, if we wanted to temporarily change a unit's team we would pass a team in as an override. If we then had another effect that overrode the team that we wanted to prioritize, we could add that new team as an override with a higher priority number.

Added some new functions to GMath to simplify angle operations.

Added a new shove ability.

Changed the naming scheme on ability to now use the format TREENAME_ABILITYNAME.

Changed the Bandit tree to the Roguery tree.

Added 2 new selection types: MultiSelectionType and RadialSelection.
MultiSelectionType allows multiple selection types to be chained together with data being passed between them for initialization.
RadialSelection allows for an optionally constrained direction and magnitude to be selected in relation to a source tile.

Added MoveContracts. These contain information pertaining to unit movement, including all tiles in a path, the animation that should be played when moving, what kind of movement type is being used, whether the path is a full path or a partial path, etc. MoveContracts are created via helper functions in the MovementHelper static class.

Used the new perlin map implementation to add variable heights to each tile.

Added a new step to the ResourceManager that bounds any resource set operations for certain resources. Ie, resources like stamina can no longer go below 0 and resources like max stamina can no longer go below 1.

Added a lock timeout to the Window.QueueToRenderCycle set up. This should prevent any infinite stalls when encountering a circular set of locks.