Added names and descriptions for several newer abilities/items.
Reworked text handling slightly. Added a clear color field to text to specify what color of background will be blended into the text to provide anti-aliasing. This field defaults to white. The clear color is then removed from the text bitmap before loading the texture (the text's edges will look slightly like the blend color so ideally the blend color should be close to the text color OR it should match the background it will be displayed in front of)
Changed the scaling of text to fix the faint blur that was present before. Additionally, changed the text scaling to now be scaled correctly at a scale of (1, 1). Now, passing different font sizes will directly affect the size of the text instead of requiring the scaling to be manually determined by the text consumer.

Changed ChainConditions to now primarily use a function instead of a parsed string. The purpose of the string parsing was to create handling agnostic of whichever ability we happened to be using but this was solved by abstracting all of the selection handling out of abilities.