Started removing Ledger implementations in favor of a new Logger api.
The logger api works based on an open ended tree/file-like structure where data can be stored in arbitrarily deep chains of nested Dictionary<string, object> objects. This is an improvement over the previous Ledger api since it allows functionality to be built for each system individually while also allowing all of the data to reside in one place and be interaccessible by each module.

Added a reference to an external DataObject assembly which contains some boilerplate for automatically managing the nested Dictionary<string, object> structure in memory as well as methods to serialize the structure to disk.

Updated the script manager to have a convenient way to add and remove object references from the scripting engine.

Added handling for LoggerActions (which react to state changes in the Logger api) and LoggerPackets (which deliver data regarding state changes to the Logger api)