Fixed an issue with text objects where the bounds being used were tile bounds as opposed to square UI bounds. This was causing slight discrepancies in the size of the object which only became noticeable due to testing the input component.
Made the dev console text a little easier to look at

Changed the input component's default font to Cascadia Mono to simplify cursor placement in the string (ie kick the can down the road)

Added clipboard support to the dev console

Added a handful of dictionary extension methods to simplify scripting interactions.

Completely reimplemented quest handling with data objects in mind. A json file named tools_0 has been added to the documentation folder that contains a working quest example at id 0. To load this quest, place the data object in :static.10000 and run StartQuest(0) in the dev console.

Moved the processing of logger packets to the main thread.

Added a UnitHurt packet.

Expanded on the script capabilities of the engine a bit more including filling out the init.js file with more base definitions as well as adding several new features to the engine-specific transpiled scripting language.

Removed context switching from the main window. Since there are ImGui bindings for C#, I will be going that route if in-engine tools become necessary.