Implemented a new method of text rendering whereby glyphs are generated using the FreeType library and text is batch rendered using glDrawArraysInstanced to index a glyph buffer that stores all glyph vertex information for every loaded font.
Reworked how UI scissoring was handled to now use the stencil buffer (which allows for much more nuanced handling especially when dealing with nested scissor areas).

Removed old TextEntry handling and replaced it with the new data object style TextEntry class.

Abstracted the transformation matrix base class out more and changed RenderableObject to now inherit more of its functionality instead of implementing it directly.

Renamed old System.Drawing text implementation. This implementation will be replaced by the new FreeType text handling shortly.

Added a RenderBatch class which I plan to use when implementing new render functions going forward. It accepts a list of any objects (which greatly simplifies handling with a very minor performance hit as casting from object -> class is quite cheap) and the types will be known by the render function called so type safety should not be an issue.

Added a general purpose array buffer to the renderer. This buffer is allocated with 500kb of memory and the intention is to use glBufferSubData to fill the buffer with whatever might be necessary for any specific render operation. That said, new render methods should try to create their own buffers to reduce data transfers if feasible.

Removed the excessive amount of typed RenderObjectInstancedGeneric functions and removed the scissor handling from the base function as it is now handled upstream. The hope for the future is that some handling can be trimmed down further with the addition of texture arrays (potentially allowing us to remove the recursive call list functionality completely).

Greatly improved the handling surrounding queuing UI to render.

Added support to the Shader class for geometry shaders.

Added a DataObjectChanged logger event

Changed the CreateObjectDefinition handling to always use a static quad definition when assigning vertices. When dealing with a large amount of quad objects this should reduce the memory cost by a fair bit.

Added new text and scissor shaders.

Added the current OS to WindowConstants.