Added the basic layout for the new UI system.
The idea is that UIElements will be positioned solely by Layout elements with Layout elements being constructed such that walking the entire UI tree won't be necessary to update render or Z position data in every circumstance.
This greatly simplifies the creation of UI by moving the responsibility of placing UI elements on the screen to the layout components.

What remains to be added in the first round of changes:
-Current UIObject handling to UIElement (positioning, dimension calculations, etc)
-Text handling (I'm currently undecided if this should be handled as a special case of the Visual class or if it should be handled solely by Label/Input/TextBox/etc components)
-Bounds checking (most of this is already present in BaseObject handling so it can be lifted from there)
-Creation of RenderBatches based on render types (sorting of opaque vs transparent types, textured types, etc)
-The UI shaders responsible for rendering these batches